convert prototype node-sdk to typescript
