+++ node/package.json
@@ -7,6 +7,7 @@
     "node": "^10.13.0 || ^12.13.0"
   },
   "scripts": {
+    "compile": "tsc",
     "test": "echo 'No unit tests (yet)'"
   },
   "author": "",
@@ -18,5 +19,10 @@
     "fabric-protos": "^2.2.2",
     "jsrsasign": "^9.1.3",
     "protobufjs": "^6.10.1"
+  },
+  "devDependencies": {
+    "@types/elliptic": "^6.4.12",
+    "@types/jsrsasign": "^8.0.7",
+    "typescript": "^4.0.3"
   }
 }
