+++ node/tsconfig.json
@@ -1,0 +1,20 @@
+{
+	"compilerOptions": {
+		"allowJs": false,
+		"declaration": true,
+		"alwaysStrict": true,
+		"module": "commonjs",
+		"outDir": "./lib",
+		"rootDir": "./src",
+		"sourceMap": true,
+		"strict": true,
+		"target": "es2017",
+		"forceConsistentCasingInFileNames": true
+	},
+	"include": [
+		"src/"
+	],
+	"exclude": [
+		"node_modules/"
+	]
+}
